refactor(engine): split the four files the builders pushed over the limit
Adding the `with_*` builders took `lifecycle/mod.rs` to 603 code lines,
`query/mod.rs` to 520, `build/mod.rs` to 518 and `build/pull.rs` to 510,
against a hard limit of 500. The builders only exposed the problem: all four
were already between 441 and 494 before this branch, so trimming just enough
to pass would have left them a single change away from breaking again.

Each file splits along a seam that already existed:

* `lifecycle/options.rs` takes `RunOptions`/`RunOverrides` — published
  surface, not behaviour — and `lifecycle/teardown.rs` takes `down` and the
  image removal it performs. Startup and teardown share nothing beyond the
  level order they each invert.
* `query/options.rs` takes the `images`/`logs` option structs. The free
  helpers below them stay: they are log plumbing, not surface.
* `build/pull_tests.rs` and `build/build_tests.rs` take the inline test
  modules, following the `#[path]` convention `pull_typo_tests.rs` already
  established here for exactly this reason.

No behaviour changes and no signature changes. The four files now sit at
355, 424, 428 and 259, and nothing under `internal/` exceeds 500.
