+++ internal/engine/lifecycle/mod.rs
@@ -8,6 +8,7 @@ mod images;
 // (#1219) fans out against the same `MAX_LIFECYCLE_CONCURRENCY` ceiling rather
 // than growing a second concurrency limit that could silently drift from this
 // one.
+mod options;
 pub(in crate::engine) mod parallel;
 mod prefetch;
 mod readiness;
@@ -18,6 +19,7 @@ mod schedule;
 mod seed;
 mod signal;
 mod targets;
+mod teardown;
 
 use std::collections::{HashMap, HashSet};
 
@@ -27,214 +29,14 @@ use crate::libpod::API_PREFIX;
 
 use readiness::SharedReady;
 
+pub use options::{RunOptions, RunOverrides};
 pub use targets::validate_stop_timeout;
 use targets::{expand_targets, filter_services, in_started_set, validate_targets};
 
 use super::container::config_hash;
 
-use super::network::resolve_network_name;
 use super::profiles::{active_profiles_set, enabled_profile_services};
 use super::Engine;
-
-/// Options for [`Engine::run`].
-///
-/// `#[non_exhaustive]` since 3.7.2, so a new field can be added in a minor
-/// release without breaking every external caller that built the struct with
-/// a literal. Construct it via [`RunOptions::new`] or the `with_*` builders
-/// below; a struct literal is refused outside this crate, which is what buys
-/// the room to grow.
-#[derive(Default)]
-#[non_exhaustive]
-pub struct RunOptions {
-	/// Override the default service command.
-	pub cmd: Vec<String>,
-	/// Remove the container after it exits.
-	pub rm: bool,
-	/// Start the container in the background without streaming logs.
-	pub detach: bool,
-	/// Additional environment variables (`KEY=VAL` strings, override service env).
-	pub env_overrides: Vec<String>,
-	/// Override the generated container name.
-	pub name_override: Option<String>,
-	/// Publish the service's declared `ports:` (compose `run --service-ports`).
-	/// When false, `run` leaves ports unpublished to avoid host-port collisions.
-	pub service_ports: bool,
-}
-
-impl RunOptions {
-	/// Every `docker compose run` option that lives on the published struct,
-	/// in CLI order. A constructor rather than a struct literal because the
-	/// type is `#[non_exhaustive]`, so the next field to land is not a
-	/// breaking change for anyone building one.
-	#[allow(clippy::too_many_arguments)]
-	pub fn new(
-		cmd: Vec<String>,
-		rm: bool,
-		detach: bool,
-		env_overrides: Vec<String>,
-		name_override: Option<String>,
-		service_ports: bool,
-	) -> Self {
-		Self {
-			cmd,
-			rm,
-			detach,
-			env_overrides,
-			name_override,
-			service_ports,
-		}
-	}
-
-	/// Override the default service command. Builder-style.
-	#[must_use]
-	pub fn with_cmd(mut self, cmd: Vec<String>) -> Self {
-		self.cmd = cmd;
-		self
-	}
-
-	/// Remove the container after it exits, `--rm`. Builder-style.
-	#[must_use]
-	pub fn with_rm(mut self, rm: bool) -> Self {
-		self.rm = rm;
-		self
-	}
-
-	/// Start the container in the background without streaming logs, `-d/--detach`.
-	/// Builder-style.
-	#[must_use]
-	pub fn with_detach(mut self, detach: bool) -> Self {
-		self.detach = detach;
-		self
-	}
-
-	/// Additional environment variables (`KEY=VAL` strings, override service env).
-	/// Builder-style.
-	#[must_use]
-	pub fn with_env_overrides(mut self, env_overrides: Vec<String>) -> Self {
-		self.env_overrides = env_overrides;
-		self
-	}
-
-	/// Override the generated container name. Builder-style.
-	#[must_use]
-	pub fn with_name_override(mut self, name_override: Option<String>) -> Self {
-		self.name_override = name_override;
-		self
-	}
-
-	/// Publish the service's declared `ports:` (compose `run --service-ports`).
-	/// When false, `run` leaves ports unpublished to avoid host-port collisions.
-	/// Builder-style.
-	#[must_use]
-	pub fn with_service_ports(mut self, service_ports: bool) -> Self {
-		self.service_ports = service_ports;
-		self
-	}
-}
-
-/// Extra `docker compose run` flag overrides threaded through the engine
-/// builder ([`Engine::with_run_overrides`]).
-///
-/// `#[non_exhaustive]` since 3.7.2, same rationale as [`RunOptions`]: the
-/// next flag to land is not a breaking change for anyone building one with a
-/// builder or constructor.
-#[derive(Default, Clone)]
-#[non_exhaustive]
-pub struct RunOverrides {
-	/// Run the command as this user (`-u/--user`, `name or UID[:GID]`).
-	pub user: Option<String>,
-	/// Working directory inside the container (`-w/--workdir`).
-	pub workdir: Option<String>,
-	/// Override the image entrypoint (`--entrypoint`).
-	pub entrypoint: Option<String>,
-	/// Extra ad-hoc volume mounts in compose short form (`-v/--volume`).
-	pub volumes: Vec<String>,
-	/// Extra published ports in compose short form (`-p/--publish`).
-	pub publish: Vec<String>,
-	/// Keep STDIN open on the container (`-i/--interactive`).
-	pub interactive: bool,
-	/// Do not start `depends_on` services before the run (`--no-deps`).
-	pub no_deps: bool,
-}
-
-impl RunOverrides {
-	/// Every `docker compose run` override, in CLI order. A constructor rather
-	/// than a struct literal because the type is `#[non_exhaustive]`, so the
-	/// next flag to land is not a breaking change for anyone building one.
-	#[allow(clippy::too_many_arguments)]
-	pub fn new(
-		user: Option<String>,
-		workdir: Option<String>,
-		entrypoint: Option<String>,
-		volumes: Vec<String>,
-		publish: Vec<String>,
-		interactive: bool,
-		no_deps: bool,
-	) -> Self {
-		Self {
-			user,
-			workdir,
-			entrypoint,
-			volumes,
-			publish,
-			interactive,
-			no_deps,
-		}
-	}
-
-	/// Run the command as this user (`-u/--user`, `name or UID[:GID]`).
-	/// Builder-style.
-	#[must_use]
-	pub fn with_user(mut self, user: Option<String>) -> Self {
-		self.user = user;
-		self
-	}
-
-	/// Working directory inside the container (`-w/--workdir`). Builder-style.
-	#[must_use]
-	pub fn with_workdir(mut self, workdir: Option<String>) -> Self {
-		self.workdir = workdir;
-		self
-	}
-
-	/// Override the image entrypoint (`--entrypoint`). Builder-style.
-	#[must_use]
-	pub fn with_entrypoint(mut self, entrypoint: Option<String>) -> Self {
-		self.entrypoint = entrypoint;
-		self
-	}
-
-	/// Extra ad-hoc volume mounts in compose short form (`-v/--volume`).
-	/// Builder-style.
-	#[must_use]
-	pub fn with_volumes(mut self, volumes: Vec<String>) -> Self {
-		self.volumes = volumes;
-		self
-	}
-
-	/// Extra published ports in compose short form (`-p/--publish`).
-	/// Builder-style.
-	#[must_use]
-	pub fn with_publish(mut self, publish: Vec<String>) -> Self {
-		self.publish = publish;
-		self
-	}
-
-	/// Keep STDIN open on the container (`-i/--interactive`). Builder-style.
-	#[must_use]
-	pub fn with_interactive(mut self, interactive: bool) -> Self {
-		self.interactive = interactive;
-		self
-	}
-
-	/// Do not start `depends_on` services before the run (`--no-deps`).
-	/// Builder-style.
-	#[must_use]
-	pub fn with_no_deps(mut self, no_deps: bool) -> Self {
-		self.no_deps = no_deps;
-		self
-	}
-}
 
 impl Engine {
 	/// Start all services defined in the compose file, creating containers that do not exist.
@@ -702,237 +504,6 @@ impl Engine {
 		}
 
 		Ok(())
-	}
-
-	/// Stop and remove all containers for the project. Does not remove volumes unless `remove_volumes` is set.
-	pub async fn down(&self, file: &ComposeFile) -> Result<()> {
-		self.down_with_options(file, false).await
-	}
-
-	/// Stop and remove services in reverse dependency order. Optionally removes named volumes and orphaned containers.
-	pub async fn down_with_options(&self, file: &ComposeFile, remove_volumes: bool) -> Result<()> {
-		let mut levels = crate::compose::resolve_levels(file)?;
-		// Teardown inverts startup: a dependent must stop before the service it
-		// depends on, so the dependency levels `up` would walk front-to-back are
-		// walked back-to-front here (the same inversion the other lifecycle
-		// commands' level walk uses, see `parallel.rs`).
-		levels.reverse();
-
-		// Prefetch every project container once and group by service, instead of
-		// one container-list round-trip per service (S+1 → 1 for the level walk).
-		let live_by_service = self.live_project_replicas().await?;
-
-		// Seed from the containers Podman actually has, walked in the same
-		// reversed level order the teardown below uses, so the board predicts
-		// what will happen rather than what the file describes. A service in the
-		// file that was never created must not sit on the board as a row that
-		// never moves.
-		let live_order: Vec<String> = levels
-			.iter()
-			.flatten()
-			.filter_map(|svc| live_by_service.get(svc))
-			.flatten()
-			.cloned()
-			.collect();
-		crate::ui::progress::begin(self.down_resources(file, &live_order, remove_volumes));
-
-		// Best-effort across every level/container/network/volume so one failure
-		// never leaves the rest of the teardown undone, but the first real
-		// REMOVAL failure is remembered and returned at the end instead of being
-		// swallowed into a warning — a `down` whose container/network/volume
-		// removal genuinely fails (storage error, active exec session) must exit
-		// non-zero, not print a warning and exit 0 (#598). A stalled or failed
-		// `stop` does NOT count towards this: the force-remove below SIGKILLs the
-		// container regardless (see `container_rm_path`), so only the removal
-		// outcome is aggregated. A 404 (already gone) stays an idempotent no-op
-		// throughout.
-		//
-		// Levels are walked strictly in order — every container in one level is
-		// attempted before the next level starts, preserving the dependency
-		// inversion above — but the containers *within* one level tear down
-		// concurrently via `join_bounded`, which returns results in input
-		// (service, then container) order rather than completion order. That
-		// keeps "the first error" deterministic regardless of which container
-		// happens to finish first: `first_error` picks the earliest in that
-		// fixed order, and since levels themselves are visited in a fixed
-		// sequence, only the first level with any failure can ever set
-		// `first_err` — a later level's failure is never mistaken for "first".
-		let mut first_err: Option<crate::error::ComposeError> = None;
-
-		for level in &levels {
-			let futs = level.iter().flat_map(|name| {
-				let service = &file.services[name];
-				let grace = self.grace_period_secs(service);
-				// Act only on containers Podman actually has. A defined-but-never-
-				// created service (or one already torn down) has no live
-				// containers, so it contributes nothing here rather than
-				// synthesizing predicted names and POSTing stop/rm to them —
-				// those 404 and, pre-fix, leaked warnings. docker compose
-				// enumerates by label and treats "nothing there" as a quiet
-				// idempotent no-op (#758).
-				live_by_service
-					.get(name)
-					.filter(|live| !live.is_empty())
-					.into_iter()
-					.flatten()
-					.map(move |container_name| {
-						self.teardown_one_container(
-							container_name,
-							grace,
-							&service.pre_stop,
-							remove_volumes,
-						)
-					})
-			});
-			if let Some(e) = parallel::first_error(parallel::join_bounded(futs).await) {
-				first_err.get_or_insert(e);
-			}
-		}
-
-		// Scaled replicas (`up --scale`/`scale`) carry the `podup.service` label
-		// of a service still in the file, so the level walk above already swept
-		// them via `live_by_service`. Orphan containers of services *removed* from
-		// the file are deliberately NOT touched here: docker compose only reaps
-		// them under `--remove-orphans`, which the dispatch layer handles via
-		// `remove_orphans` before teardown. Removing them unconditionally here
-		// made that flag a no-op.
-
-		for (key, config) in &file.networks {
-			let external = config.as_ref().and_then(|c| c.external).unwrap_or(false);
-			if external {
-				continue;
-			}
-			let network_name = resolve_network_name(key, file, &self.project);
-			let net_path = format!(
-				"{API_PREFIX}/networks/{}",
-				crate::libpod::urlencoded(&network_name),
-			);
-			// `delete_existed`, not `delete_ok`: this loop walks the networks the
-			// compose file *declares*, which is not the same set as the networks
-			// that exist. `delete_ok` throws away the boolean that tells the two
-			// apart, so every 404 arrived here as `Ok(())` and was announced as a
-			// removal — measured on a project that had never been created,
-			// `down -v` reported removing two networks and a volume, none of
-			// which had ever existed. The `Err(404)` arm below was unreachable
-			// for the same reason: the layer underneath had already turned the
-			// 404 into a success.
-			crate::ui::progress::start("Network", &network_name, "Removing");
-			match self.client.delete_existed(&net_path).await {
-				Ok(true) => crate::ui::progress_line("Network", &network_name, "Removed"),
-				// The network was already gone (404) — nothing to do, but the row
-				// has to close, or the live board leaves it spinning on
-				// `Removing` forever (#1347).
-				Ok(false) => crate::ui::progress_line("Network", &network_name, "Absent"),
-				Err(e) => {
-					tracing::warn!("could not remove network {network_name}: {e}");
-					// Close the row visibly — a `down` whose removal genuinely
-					// failed previously hid the failure behind a spinner (#1347).
-					crate::ui::progress_line("Network", &network_name, "Failed");
-					first_err.get_or_insert(crate::error::ComposeError::Podman(e));
-				}
-			}
-		}
-
-		// Sweep any remaining project networks by label — the implicit
-		// `<project>_default` (present only when the file was normalized), or a
-		// network whose compose key changed — mirroring the container sweep so
-		// teardown is complete regardless of how the file was parsed. Only
-		// podup-labelled networks match, so external networks are never touched.
-		// This is a supplementary catch-all on top of the file-driven network
-		// loop above (which already aggregates its own failures into
-		// `first_err`), so a failure here is intentionally swallowed rather than
-		// folded in again.
-		let _ = self.remove_project_networks_by_label().await;
-
-		if remove_volumes {
-			for (key, config) in &file.volumes {
-				let external = config.as_ref().and_then(|c| c.external).unwrap_or(false);
-				if external {
-					continue;
-				}
-				let volume_name = config
-					.as_ref()
-					.and_then(|c| c.name.as_deref())
-					.map(|s| s.to_string())
-					.unwrap_or_else(|| format!("{}_{}", self.project, key));
-				let vol_path = format!(
-					"{API_PREFIX}/volumes/{}",
-					crate::libpod::urlencoded(&volume_name),
-				);
-				// See the network loop above: only a delete that found something
-				// may be reported, and a volume is the object where a false
-				// "Removed" is worst — it names data the operator believes is
-				// gone.
-				crate::ui::progress::start("Volume", &volume_name, "Removing");
-				match self.client.delete_existed(&vol_path).await {
-					Ok(true) => crate::ui::progress_line("Volume", &volume_name, "Removed"),
-					// The volume was already gone (404) — nothing to do, but the
-					// row has to close, or the live board leaves it spinning on
-					// `Removing` forever (#1347).
-					Ok(false) => crate::ui::progress_line("Volume", &volume_name, "Absent"),
-					Err(e) => {
-						tracing::warn!("could not remove volume {volume_name}: {e}");
-						// A `down -v` whose volume removal genuinely failed
-						// previously hid the failure behind a spinner (#1347).
-						crate::ui::progress_line("Volume", &volume_name, "Failed");
-						first_err.get_or_insert(crate::error::ComposeError::Podman(e));
-					}
-				}
-			}
-		}
-
-		// Internal native secrets are podup-owned (not user data), so remove
-		// them unconditionally — independent of `remove_volumes`.
-		let secrets = self.remove_internal_secrets(file).await;
-
-		// Close the board before returning, on every path: the region hides the
-		// cursor and an early `?` would leave the terminal without one.
-		crate::ui::progress::end();
-		secrets?;
-
-		if let Some(e) = first_err {
-			return Err(e);
-		}
-		Ok(())
-	}
-
-	/// Remove the images used by the project's services (`down --rmi`). With
-	/// `local_only`, only images of services that build locally (a `build:`
-	/// section) are removed — matching `docker compose down --rmi local`.
-	pub async fn remove_service_images(&self, file: &ComposeFile, local_only: bool) -> Result<()> {
-		// Aggregate like every sibling loop in this teardown: complete the sweep,
-		// then report the first real failure. Warning and returning Ok meant
-		// `down --rmi` exited 0 having left images behind.
-		let mut first_err: Option<crate::error::ComposeError> = None;
-		for (name, service) in &file.services {
-			let builds_locally = service.build.is_some();
-			if local_only && !builds_locally {
-				continue;
-			}
-			let image = match &service.image {
-				Some(img) => img.clone(),
-				None if builds_locally => self.service_image_tag(name, service),
-				None => continue,
-			};
-			// Do NOT force: a force-removal cascades to every container using the
-			// image — including ones owned by other compose projects that share it
-			// (e.g. two stacks both on `nginx:latest`). docker compose leaves an
-			// in-use image in place, so an "in use" conflict is a skip, not a
-			// failure.
-			let path = format!("{API_PREFIX}/images/{}", crate::libpod::urlencoded(&image),);
-			match self.client.delete_ok(&path).await {
-				Ok(_) => crate::ui::progress_line("Image", &image, "Removed"),
-				Err(e) if e.is_status(404) => {}
-				Err(e) if e.is_image_in_use() => {
-					tracing::debug!("image {image} is still in use — skipping removal")
-				}
-				Err(e) => {
-					tracing::warn!("could not remove image {image}: {e}");
-					first_err.get_or_insert(crate::error::ComposeError::Podman(e));
-				}
-			}
-		}
-		first_err.map_or(Ok(()), Err)
 	}
 }
